The excel module code works
The excel module code works, the table is created with the necessary fields, but there are disadvantages. First, if the field with the average value is set to position 4 (by default) when creating the table (excel initialization), then data is layered (fields with average values replace the fields of months), second, if the field with the average value is set to other positions during initialization, then they are replaced by the value of months, mess and third, if the field with the average value is not set, it works as it should, the column appears last, after all months, the average values are calculated correctly, but, the number of columns with an average value directly depends on the number of changes made to the table by the user.